+++ packages/cx-studio/cx_studio/path_expander/__init__.py
@@ -4,3 +4,4 @@ from .cx_executablevalidator import *
 from .cx_pathexpander import *
 from .cx_pathvalidator import *
 from .cx_suffixvalidator import *
+from .cx_suffix_finder import *